refactor: move common functions to test_util
